mesh indices set to u16:
  No mesh needs more than 65536 indices. If so, we split into multiple meshes.
  This saves graphics memory.

+++ src/mesh.rs
@@ -13,7 +13,7 @@ pub struct Mesh<V: Vertex> {
 
     /// Indices of vertices for forming triangles
     /// Meshes currently use indexed draw only
-    pub indices: Vec<(u32,u32,u32)>,
+    pub indices: Vec<(u16,u16,u16)>,
 
     /// String values
     /// Can use to store names of textures, e.g.: ("bump_texture", "mybump.obj")
